Add some soap fault code.

+++ rt/transports/jms-spec/src/main/java/org/apache/cxf/transport/jms/uri/JMSEndpoint.java
@@ -80,7 +80,7 @@ public class JMSEndpoint {
         return deliveryMode;
     }
 
-    public int getTimeToLive() {
+    public long getTimeToLive() {
         String timeToLive = getParameter(JMSURIConstants.TIMETOLIVE_PARAMETER_NAME);
         if (timeToLive == null) {
             return JMSURIConstants.TIMETOLIVE_DEFAULT;
